backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
chromadb
pdfplumber
google-generativeai
python-dotenv
